# Reduce locks