@LilithSilver 2023-10-20
Added one last thing -- removed a failing test that was debug-only since it's no longer relevant (I removed strict is-flushed checking because we can no longer be sure it actually works (flushing a dependency now flushes all dependents as well, and there's no way to crawl up the chain and check without iterating through all jobs). If we really want we can add a MarkFlushed() thing on `Job` but that's a job for a later PR.